+++ examples/minimal/providers.tf
@@ -1,2 +1,3 @@
-provider "null" {}
-
+provider "azurerm" {
+  version         = "=2.11.0"
+}