+++ CHANGELOG.md
@@ -1,3 +1,5 @@
+# Tamr Azure HDInsights HBase module - v0.1.2 - June 1st 2020
+* Updates the module to be compliant with Azure provider 2.11.0
 # Tamr Azure HDInsights HBase module - v0.1.1 - May 27th 2020
 * Explicitly sets enable_https_traffic_only to true
 # Tamr Azure HDInsights HBase module - v0.1.0 - May 12th 2020